Quiz: show "Not answered" for missing answers
