+++ src/app/components/content/QuizQuestion.tsx
@@ -9,27 +9,29 @@ import {isDefined} from "../../services/miscUtils";
 import {SITE, SITE_SUBJECT} from "../../services/siteConstants";
 import {IsaacLinkHints, IsaacTabbedHints} from "./IsaacHints";
 import {IsaacContent} from "./IsaacContent";
+import {QuizAttemptDTO} from "../../../IsaacApiTypes";
 
-export const QuizAttemptContext = React.createContext<{quizAttemptId: number | null}>({quizAttemptId: null});
+export const QuizAttemptContext = React.createContext<{quizAttempt: QuizAttemptDTO | null}>({quizAttempt: null});
 
 export const QuizQuestion = ({doc}: { doc: ApiTypes.IsaacQuestionBaseDTO }) => {
     const dispatch = useDispatch();
 
-    const {quizAttemptId} = useContext(QuizAttemptContext);
+    const {quizAttempt} = useContext(QuizAttemptContext);
 
     useEffect((): (() => void) => {
         return () => {
             // Submit answer when unmounting if it is dirty
-            if (isDefined(quizAttemptId)) {
-                dispatch(submitQuizQuestionIfDirty(quizAttemptId, doc.id as string));
+            if (isDefined(quizAttempt) && quizAttempt.completedDate === undefined) {
+                dispatch(submitQuizQuestionIfDirty(quizAttempt.id as number, doc.id as string));
             }
         };
-    }, [dispatch, doc.id, quizAttemptId]);
+    }, [dispatch, doc.id, quizAttempt]);
 
     const validationResponse = doc?.bestAttempt;
-    const validated = validationResponse?.correct !== undefined;
+    const validated = validationResponse?.correct !== undefined || (!validationResponse && isDefined(quizAttempt?.completedDate));
     const correct = validationResponse?.correct;
     const sigFigsError = (validationResponse?.explanation?.tags || []).includes("sig_figs");
+    const noAnswer = validated && correct === undefined;
 
     const QuestionComponent = QUESTION_TYPES.get(doc.type || "default");
 
@@ -46,11 +48,11 @@ export const QuizQuestion = ({doc}: { doc: ApiTypes.IsaacQuestionBaseDTO }) => {
             </React.Fragment>}
 
             {/* Validation Response */}
-            {validationResponse && validated && <div className={`validation-response-panel p-3 mt-3 ${correct ? "correct" : ""}`}>
+            {validated && <div className={`validation-response-panel p-2 mt-2 ${correct ? "correct" : ""}`}>
                 <div className="pb-1">
-                    <h1 className="m-0">{sigFigsError ? "Significant Figures" : correct ? "Correct!" : "Incorrect"}</h1>
+                    <h3 className="m-0">{noAnswer ? "Not answered" : sigFigsError ? "Significant Figures" : correct ? "Correct!" : "Incorrect"}</h3>
                 </div>
-                {validationResponse.explanation && <div className="mb-2">
+                {validationResponse && validationResponse.explanation && <div className="mb-1">
                     <IsaacContent doc={validationResponse.explanation} />
                 </div>}
             </div>}
